Move registration code to Program
